simpleapp/requirements.txt
Flask==0.12.1
Flask-PyMongo==0.4.1
Jinja2==2.9.6
MarkupSafe==1.0
Werkzeug==0.12.1
click==6.7
future==0.16.0
itsdangerous==0.24
pymongo==3.4.0
python-dateutil==2.6.0
six==1.10.0
wsgiref==0.1.2
